refactor creation and deletion invoice forms actions

+++ lib/actions.ts
@@ -3,11 +3,11 @@
 import { FormSchema, LoginSchema } from '@/lib/schemas'
 import { sql } from '@/lib/db-connection'
 import { revalidatePath } from 'next/cache'
-import { signIn } from '@/auth/auth'
-import { isUserLogged } from '@/auth/logged-user'
+import { auth, signIn } from '@/auth/auth'
 import { AuthError } from 'next-auth'
 import { isRedirectError } from 'next/dist/client/components/redirect'
 import { validateFormData } from '@/lib/schema-validation'
+// import { authAccessControl } from '@/auth/auth-provider'
 
 // hydration error
 // const dateToDatabase = () => new Date().toISOString().split('T')[0]
@@ -20,10 +20,14 @@ export type State = {
   }
   message?: string | null
 }
+type ExtendedState = State & { success?: string; error?: string }
 
 const CreateInvoice = FormSchema.omit({ id: true, date: true })
-export async function createInvoice(formData: FormData) {
-  isUserLogged()
+
+export async function createInvoice(
+  formData: FormData
+): Promise<ExtendedState> {
+  if (!(await auth())) return { error: 'Log in to update an invoice.' }
 
   const validatedFields = validateFormData(CreateInvoice, formData, {
     errorMessage: 'Missing Fields. Failed to Create Invoice.',
@@ -48,8 +52,11 @@ export async function createInvoice(formData: FormData) {
 }
 
 const UpdateInvoice = FormSchema.omit({ id: true, date: true })
-export async function updateInvoice(formData: FormData, id: string) {
-  isUserLogged()
+export async function updateInvoice(
+  formData: FormData,
+  id: string
+): Promise<ExtendedState> {
+  if (!(await auth())) return { error: 'Log in to update an invoice.' }
 
   const validatedFields = validateFormData(UpdateInvoice, formData, {
     errorMessage: 'Missing Fields. Failed to Update Invoice.',
@@ -76,7 +83,7 @@ export async function updateInvoice(formData: FormData, id: string) {
 }
 
 export async function deleteInvoice(id: string) {
-  isUserLogged()
+  if (!(await auth())) return { error: 'Log in to delete an invoice.' }
   try {
     await sql`
     DELETE FROM invoices
